# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: City Stories Guide API
  description: 'REST API for City Stories Guide — an ambient audio guide app that tells stories

    about nearby places while walking. Provides endpoints for discovering nearby stories,

    managing cities/POIs/stories, authentication, purchases, and admin operations.

    '
  version: 1.0.0
  contact:
    name: City Stories Guide Team
servers:
- url: /api/v1
  description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from auth endpoints
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          example: invalid request
        trace_id:
          type: string
          description: Request trace ID for debugging
          example: abc-123-def
    ValidationDetail:
      type: object
      required:
      - field
      - message
      properties:
        field:
          type: string
          description: JSON field name that failed validation
          example: email
        message:
          type: string
          description: Human-readable validation failure reason
          example: must be a valid email address
    ValidationError:
      type: object
      required:
      - error
      - details
      properties:
        error:
          type: string
          enum:
          - validation_error
          example: validation_error
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationDetail'
        trace_id:
          type: string
          description: Request trace ID for debugging
          example: abc-123-def
    ErrorOrValidationError:
      oneOf:
      - $ref: '#/components/schemas/Error'
      - $ref: '#/components/schemas/ValidationError'
...
    CursorPaginatedResponse:
      description: Common cursor-based pagination response fields
      type: object
      required:
      - items
      - next_cursor
      - has_more
      properties:
        items:
          type: array
          description: Page items. Concrete endpoints specify the item schema.
          items: {}
        next_cursor:
          type: string
          description: Opaque cursor to pass as cursor param for the next page. Empty string if no more pages.
          example: aWQ6NDI=
        has_more:
          type: boolean
          description: Whether there are more pages after this one
          example: true
    MessageResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
...
    CityResponse:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/City'
...
    DeviceTokenResponse:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/DeviceToken'
...
    CitiesPageResponse:
      allOf:
      - $ref: '#/components/schemas/CursorPaginatedResponse'
      - type: object
        properties:
          items:
            type: array
            items:
              $ref: '#/components/schemas/City'
...
    RegisterDeviceTokenRequest:
      type: object
      required:
      - user_id
      - token
      - platform
      properties:
        user_id:
          type: string
          format: uuid
        token:
          type: string
          maxLength: 500
        platform:
          type: string
          enum:
          - ios
          - android
    UnregisterDeviceTokenRequest:
      type: object
      required:
      - token
      properties:
        token:
          type: string
...
    City:
      type: object
      required:
      - id
      - name
      - country
      - center_lat
      - center_lng
      - radius_km
      - is_active
      - download_size_mb
      - created_at
      - updated_at
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Tbilisi
        name_ru:
          type: string
          nullable: true
          example: Тбилиси
        country:
          type: string
          example: Georgia
        center_lat:
          type: number
          format: double
          example: 41.7151
        center_lng:
          type: number
          format: double
          example: 44.8271
        radius_km:
          type: number
          format: double
          example: 15.0
        is_active:
          type: boolean
          example: true
        download_size_mb:
          type: number
          format: double
          example: 120.5
        deleted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    DeviceToken:
      type: object
      required:
      - id
      - user_id
      - token
      - platform
      - is_active
      - created_at
      - updated_at
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          format: uuid
        token:
          type: string
          example: fcm-token-abc123
        platform:
          type: string
          enum:
          - ios
          - android
          example: ios
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /cities:
    get:
      tags:
      - Cities
      summary: List cities
      operationId: listCities
      parameters:
      - name: cursor
        in: query
        description: Opaque cursor from previous response's next_cursor
        schema:
          type: string
      - name: limit
        in: query
        description: Number of items per page (default 20, max 100)
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: Paginated list of cities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitiesPageResponse'
              example:
                items:
                  - id: 1
                    name: Tbilisi
                    name_ru: Тбилиси
                    country: Georgia
                    center_lat: 41.7151
                    center_lng: 44.8271
                    radius_km: 15.0
                    is_active: true
                    download_size_mb: 120.5
                    deleted_at: null
                    created_at: "2025-01-15T10:00:00Z"
                    updated_at: "2025-01-15T10:00:00Z"
                next_cursor: "aWQ6Mg=="
                has_more: true
        '400':
          description: Invalid cursor or limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "invalid cursor"
                trace_id: "req-abc-123"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cities/{id}:
    get:
      tags:
      - Cities
      summary: Get city by ID
      operationId: getCity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '400':
          description: Invalid city ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: City details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
        '404':
          description: City not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /device-tokens:
    post:
      tags:
      - Device Tokens
      summary: Register a device token for push notifications
      operationId: registerDeviceToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceTokenRequest'
      responses:
        '201':
          description: Device token registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTokenResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOrValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
      - Device Tokens
      summary: Unregister a device token
      operationId: unregisterDeviceToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnregisterDeviceTokenRequest'
      responses:
        '200':
          description: Device token unregistered
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/MessageResponse'
                - type: object
                  properties:
                    message:
                      example: device token unregistered
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOrValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...